Client_Server/requirements.txt
Django==1.7.7
django-cors==0.1
djangorestframework==3.1.1
pyserial==2.7
pytz==2015.2
requests==2.6.0
